requirements.txt
anthropic
ddgs
requests
python-dotenv